calculate linear regression between every two columns in a data frame. in: a dataframes. out: a dataframe showing the linear regression.

calculate linear regression between every two columns in a data frame. in: a dataframes. out: a dataframe showing the linear regression.

Usage

lmdf(data, simply = FALSE, intercept = TRUE)

Arguments

data

a dataframe

simply

logical

intercept

logical

Value

another dataframe

Examples

df <- data.frame(a = 1:10, b = 1:10 + rnorm(10), c = 1:10 + rnorm(10))
lmdf(df)
#>   x y r.squared adj.r.squared   intercept     slope Std.Error.intercept
#> 1 a b 0.9497671     0.9434880  0.05380225 1.0191131           0.5141542
#> 2 a c 0.9327415     0.9243342 -0.78092082 1.0553870           0.6217132
#> 3 b a 0.9497671     0.9434880  0.22613975 0.9319545           0.4854733
#> 4 b c 0.9122958     0.9013327 -0.62460821 0.9981253           0.7009913
#> 5 c a 0.9327415     0.9243342  1.06009256 0.8837910           0.4970337
#> 6 c b 0.9122958     0.9013327  1.06720938 0.9140092           0.5935214
#>   Std.Error.slope t.intercept   t.slope Pr.intercept     Pr.slope
#> 1      0.08286345   0.1046422 12.298705   0.91923568 1.777173e-06
#> 2      0.10019815  -1.2560789 10.532999   0.24452690 5.752794e-06
#> 3      0.07577664   0.4658129 12.298705   0.65376855 1.777173e-06
#> 4      0.10941644  -0.8910356  9.122261   0.39891469 1.677936e-05
#> 5      0.08390687   2.1328385 10.532999   0.06550284 5.752794e-06
#> 6      0.10019547   1.7980977  9.122261   0.10987197 1.677936e-05